+++ DEV-5/DEV-5/CatalogStateHandler.cs
@@ -5,10 +5,10 @@
     public class CatalogEventArgs
     {
         public string Message { get; private set;}
-        public int Sum { get; private set;}
+        public double Sum { get; private set;}
         public int Count { get; private set;}
 
-        public CatalogEventArgs(string _mes, int _sum, int _count)
+        public CatalogEventArgs(string _mes, double _sum, int _count)
         {
             Message = _mes;
             Sum = _sum;
